feat: postgres-based lock-storage
unfortantely lock-storage in wsgi-dav is not properly abstracted, so we have to
implement the postgres-lock-storage as MutableMapping. collections.abc is just
awesome, we have to implement few basic __dunder__ methods and the rest is
implemented for us.

wsgi-dav accesses the same lock-storage from multiple threads, but postgres only
locks transaction-based and will allow a second lock on the same transaction. So
we have to lock a normal lock and a postgres-lock.

The lock-storage (webdav lock) of wsgi-dav is not atomic, some of the state is
kept in the lock-manager. But since the lock-manager also has a lock (mutex), we
just have to inject the lock (mutex) of the lock-storage into the lock-manager,
to make the operation atomic.