feat: widen token column a bit
The 128 character limit implied a character limit of about 90 for file names.

This is a bit too close to the practical limit of what users *actually* do,
and if the limit is hit, LOCK calls fail with unexpected errors due to the DB
not being able to deal with the token.

Widening it to 255 chars should move the "hard limit" further away, thus
reducing the chance of errors significantly.