chore: mypy mini cleanups
Note regarding the `ignore`d super() call: mypy thinks that the
`super()` call in `ManabiS3Provider.__init__()` goes to the direct base
class `ManabiProvider` and is thus missing the required `root_folder`
parameter, but this ain't so: This calls `DAVProvider.__init__()` which
doesn't take that param.

Then we add a bit of typing sugar and make mypy shut up about missing
type information from 3rd party imports.

A few smaller issues that mypy finds in CI but not locally are addressed
as well.